Last Update
Added:
- itemsSold.php
-> Specific page for sellers to view items they have sold.

Update:
- edit-item.php & sell-item.php
-> Changed floatval back to intval

- profile_header.inc.php & userprofile.php & index.php
-> Added href for itemsSold.php

- processOffer.php
-> Fully functional

Wording Updates:
- processAmend.php
- offersReceived.php
- offersSent.php

+++ edit-item.php
@@ -213,7 +213,7 @@
 					//Price
 					if (empty($_POST['price'])) {
 							$price_err = "<p style='color:red;'>*You had a price before!</p>";
-					} else if ( (strval($_POST['price'])) != (strval(floatval($_POST['price']))) ) {
+					} else if ( (strval($_POST['price'])) != (strval(intval($_POST['price']))) ) {
 							$price_err = "<p style='color:red;'>*Please enter valid integers for price</p>";
 					} else {
 							$inputpass = ($inputpass + 1);
